+++ src/main/kotlin/com/balancemania/api/client/mediapipe/model/MediaPipeImgAnalysisResponse.kt
@@ -2,20 +2,20 @@ package com.balancemania.api.client.mediapipe.model
 
 data class MediaPipeImgAnalysisResponse(
     /** 정면 어깨 각도 */
-    val frontShoulderAngle: Float,
+    val frontShoulderAngle: Double,
 
     /** 정면 골반 각도 */
-    val frontPelvisAngle: Float,
+    val frontPelvisAngle: Double,
 
     /** 정면 무릎 각도 */
-    val frontKneeAngle: Float,
+    val frontKneeAngle: Double,
 
     /** 정면 발목 각도 */
-    val frontAnkleAngle: Float,
+    val frontAnkleAngle: Double,
 
     /** 측면 목 각도 */
-    val sideNeckAngle: Float,
+    val sideNeckAngle: Double,
 
     /** 측면 신체 각도 */
-    val sideBodyAngle: Float,
+    val sideBodyAngle: Double,
 )
